+++ packages/server/app/controllers/people_controller.ts
@@ -46,7 +46,11 @@ export default class PeopleController {
     const typeKey = queryData['typeKey']
     const notInCohort = queryData['notInCohort']
 
-    let [query, countQuery] = buildApiQuery(Person.query(), queryData, 'people', 'familyName')
+    let [query, countQuery] = await buildApiQuery({
+      modelQuery: Person.query(),
+      queryData,
+      tableName: 'people',
+    })
 
     query.apply((scopes: ExtractScopes<typeof Person>) => scopes.fullPerson())
 
@@ -56,6 +60,7 @@ export default class PeopleController {
       // DB context use sql column names
       countQuery.where('type_key', personType.key)
     }
+
     buildPeopleSort(query, queryData)
     if (notInCohort) {
       // Get complete list of people ids associated with cohort to filter them out