+++ packages/server/app/helpers/query_builder.ts
@@ -6,6 +6,19 @@ import Resource from '#models/resource'
 import Person from '#models/person'
 import { TransactionClientContract } from '@adonisjs/lucid/types/database'
 import Tag from '#models/tag'
+import { getIdsByDimensionQuery } from '#helpers/query_dimensions'
+
+const defaultSearchFields: Record<string, string[]> = {
+  cohorts: ['name'],
+  event_types: ['name'],
+  fact_types: ['name'],
+  people: ['family_name', 'given_name'],
+  person_types: ['name'],
+  resources: ['name'],
+  resource_types: ['name'],
+  tags: ['pathname'],
+  users: ['email'],
+}
 
 function getSortSettings(queryData: Record<string, any> = {}) {
   const sort = queryData['sort']
@@ -14,16 +27,28 @@ function getSortSettings(queryData: Record<string, any> = {}) {
   return [sort, order]
 }
 
-export function buildApiQuery(
-  modelQuery: any,
-  queryData: Record<string, any> = { limit: 10, offset: 0 },
-  tableName: string,
+export async function buildApiQuery({
+  modelQuery,
+  queryData = { limit: 10, offset: 0 },
+  tableName,
+}: {
+  modelQuery: any
+  queryData?: Record<string, any>
+  tableName: string
   defaultSearch?: string
-) {
+}) {
   let countQuery = db.from(tableName)
   const limit = queryData['limit']
   const offset = queryData['offset']
-  const search = queryData['search']
+  const searchQuery = queryData['search']
+  const filters = queryData['filter']
+  const dims = queryData['dim']
+  if (filters || dims) {
+    const typeTableName = string.singular(tableName) + '_types'
+    const result = await getIdsByDimensionQuery(tableName, typeTableName, filters, dims)
+    modelQuery.andWhereIn('id', result)
+    countQuery.andWhereIn('id', result)
+  }
 
   if (limit) {
     modelQuery.limit(limit)
@@ -32,11 +57,20 @@ export function buildApiQuery(
     modelQuery.offset(offset)
   }
 
-  // Add search functionality to modelQuery
-  if (typeof search === 'string' && defaultSearch) {
-    modelQuery.whereILike(defaultSearch, search)
-    countQuery.whereILike(defaultSearch, search)
-  } else if (typeof search === 'object' && search !== null) {
+  // Searching not supported for Events or Facts. Just ignore for those.
+  if (tableName !== 'events' && tableName !== 'facts') {
+    let search: any
+    // If simple string passed as search param, convert to search object filter so every
+    // request uses the same search code mechanism below
+    if (typeof searchQuery === 'string') {
+      search = {}
+      defaultSearchFields[tableName].forEach((key: string) => {
+        search[key] = searchQuery
+      })
+    } else {
+      search = searchQuery
+    }
+
     modelQuery.where((query: any) => {
       // Wrap all of this in .where() so the OR clauses generated below will all be enclosed in
       // parentheses in the generated SQL as one logical unit.